feat(doom): Add haptic feedback to doom (#102)
* feat(doom): Add haptic support to doom
* Update doom to have callbacks for when the player is damaged and when they fire a weapon
* Update doom implementation to trigger different haptic effects depending on the weapon fired and the amount of damage taken

Haptics are awesome, and with doom we have the actual source code so we can do some fun stuff :)

Build and run `main` on Box-Emu v0 which has haptics and ensure it feels good.

* fix sa

* improve maintainability

* more haptics

* add player interaction haptics and fix sa

* update readme

* Update components/doom/src/doom.cpp


---------

+++ components/shared_memory/src/shared_memory.c
@@ -69,7 +69,7 @@ void* shared_mem_allocate(const shared_mem_request_t* request) {
 }
 
 void shared_mem_clear(void) {
-    printf("Num bytes allocated: %d\n", current_offset_);
+    printf("Num bytes allocated: %d\n", (int)current_offset_);
     // TODO: Use SIMD-accelerated memset from ESP32s3 vector instructions
     memset(memory_pool_, 0, TOTAL_MEMORY_SIZE);
     current_offset_ = 0;
